edit test_virulencefinder.py
